Improve compatibility with bullets that have parallel actions on creation
Added a few hooks onto steps.  Now there's additional processing that can be handled when a step finishes and before it goes onto the next one.  This was necessary for Mutations.  Also fixed some mutation calculations, including adding support for multiple direction types with SetDirectionMutation.

Bullet is now attached to each step since they're not really usable between multiple bullets anyway due to how they're generated and how they maintain their own state.  A hook was added so certain step types can handle additional processing when a Bullet is assigned to them, such as sequences assigning the bullet to all child steps.

There are now parallel step sequences because of some example bullet patterns that were spawning bullets with multiple actions that weren't wrapped in a single action.  There's currently nothing in the official specification that indicates parallel processing to be a thing that any action can do, so for now I'll leave it as just something that bullets will generate with.  Every action in the mean time will remain a sequence.